Added primitive type- and rangechecking and autoboxing.
Assigning strings that would be truncated by mysql now results
in an exception. It's a silent fail by mysql that shouldn't silent.

Instead of throwing exceptions everytime something doesn't
match a specific type, autoboxing will be used.
Tinyint and bigint have very different ranges in mysql.
A check is made to be sure a value is within a specific range
(depending on the mysql type).

It is possible to specify which tests to run via ant now.
ant -Dtests=ColumnTest only executes the ColumnTest.
It simply uses phpunits --filter option.