Round #1 of the major second rewrite.
Refactored a lot out of Model. Separating the classes way more now.
Table is Entity unaware now, which is way better. Table is simple now.
So is Model. It only translates the fieldNames.
Slowly starting to implement foreign keys and compound foreign keys.
Replaced ImmutableArrayObject with ImmutableFixedArray. The framework
should start relying on indexed arrays instead of associative arrays.
Refactored cakePHPs Inflector to fir the rest of the code.